switch from docker to systemd

+++ scripts/update.sh
@@ -1,7 +1,31 @@
 #!/bin/bash
 
-docker compose pull
-docker compose down
 rm -f ./storage/logs/*.log
-docker compose up -d
-date '+%Y-%m-%d %H:%M:%S' > ./storage/app/update.txt
+date '+%Y-%m-%d %H:%M:%S Started' > ./storage/app/update.txt
+
+if type docker >/dev/null 2>&1; then
+    docker compose down
+fi
+
+SERVICE_NAME="p-manager"
+SETUP_SCRIPT="$(dirname "$0")/setup.sh"
+
+service_exists() {
+    systemctl list-units --full -all | grep -Fq "$SERVICE_NAME.service"
+}
+
+service_active() {
+    systemctl is-active --quiet $SERVICE_NAME
+}
+
+if service_exists; then
+    if service_active; then
+        systemctl restart $SERVICE_NAME
+    else
+        systemctl start $SERVICE_NAME
+    fi
+else
+    $SETUP_SCRIPT
+fi
+
+date '+%Y-%m-%d %H:%M:%S Done' >> ./storage/app/update.txt
